[refactor] mixedSound 모델 정리
- MixedSound파일 따로 생성
