+++ LullabyRecipe/Model/Sound.swift
@@ -15,19 +15,6 @@ struct Sound: Identifiable, Codable {
     let imageName: String
 }
 
-struct MixedSound: Identifiable, Codable, Equatable {
-    static func == (lhs: MixedSound, rhs: MixedSound) -> Bool {
-        return true
-    }
-    
-    let id: Int
-    let name: String
-    let baseSound: Sound?
-    let melodySound: Sound?
-    let naturalSound: Sound?
-    let imageName: String
-}
-
 var baseSounds = [
     Sound(id: 0,
           name: "Empty",
